basic structure for real time management of agents
this is clearly game-oriented, but the principle can be reused for
anything that works in real time with multiple agents.
A priority queue is used for keep sorted the commands according to their
executable time. The engine is the receptor of commands, invoked by the
priority queue.
Not that Units are base class of Squad, and mixins like Movable can
easily be added to the declared objects.